+++ OUDS/Core/Tokens/SemanticTokens/Sources/TypographySemanticTokens.swift
@@ -28,6 +28,29 @@ public typealias TypographyFontSizeSemanticToken = TypographyFontSizeRawToken
 /// The global design system tools uses verbs of semantic token for font line height, which is basically a raw token for font line height
 public typealias TypographyFontLineHeightSemanticToken = TypographyFontLineHeightRawToken
 
+// MARK: Semantic tokens
+
+// MARK: - Composite semantic token
+
+/// Composite semantic tokens which will wrap a combination of `TypographyCompositeRawToken` depending to viewports.
+public final class TypographyCompositeSemanticToken: NSObject {
+
+    /// For **extra-compact** and **compact** viewports
+    public let compact: TypographyCompositeRawToken
+    /// For **regular** and **medium** viewports
+    public let regular: TypographyCompositeRawToken
+
+    public init(compact: TypographyCompositeRawToken, regular: TypographyCompositeRawToken) {
+        self.compact = compact
+        self.regular = regular
+    }
+
+    public override func isEqual(_ object: Any?) -> Bool {
+        guard let object = object as? TypographyCompositeSemanticToken else { return false }
+        return self.compact == object.compact && self.regular == object.regular
+    }
+}
+
 // MARK: - Semantic tokens
 
 /// This is a group of semantic tokens for **typography**.
@@ -36,6 +59,8 @@ public typealias TypographyFontLineHeightSemanticToken = TypographyFontLineHeigh
 public protocol TypographySemanticTokens {
 
     // MARK: - Semantic token - Typography - Font - Family
+
+    // TODO: Should we suffix our tokens names by "iOS"?
 
     var fontFamily: TypographyFontFamilyRawToken { get }
     var fontFamilyDisplay: TypographyFontFamilySemanticToken { get }
@@ -46,6 +71,8 @@ public protocol TypographySemanticTokens {
 
     // MARK: - Semantic token - Typography - Font - Weight
 
+    var fontWeightDefault: TypographyFontWeightSemanticToken { get }
+    var fontWeightStrong: TypographyFontWeightSemanticToken { get }
     var fontWeightDisplay: TypographyFontWeightSemanticToken { get }
     var fontWeightHeading: TypographyFontWeightSemanticToken { get }
     var fontWeightBodyDefault: TypographyFontWeightSemanticToken { get }
@@ -54,7 +81,7 @@ public protocol TypographySemanticTokens {
     var fontWeightLabelStrong: TypographyFontWeightSemanticToken { get }
     var fontWeightCode: TypographyFontWeightSemanticToken { get }
 
-    // MARK: - Semantic token - Typography - Font - Size - Mobile
+    // MARK: - Semantic token - Typography - Font - Size - Mobile (extra-compact/compact)
 
     var fontSizeMobileDisplayLarge: TypographyFontSizeSemanticToken { get }
     var fontSizeMobileDisplayMedium: TypographyFontSizeSemanticToken { get }
@@ -66,10 +93,8 @@ public protocol TypographySemanticTokens {
     var fontSizeMobileBodyLarge: TypographyFontSizeSemanticToken { get }
     var fontSizeMobileBodyMedium: TypographyFontSizeSemanticToken { get }
     var fontSizeMobileBodySmall: TypographyFontSizeSemanticToken { get }
-    var fontSizeMobileCodeMedium: TypographyFontSizeSemanticToken { get }
-    var fontSizeMobileCodeSmall: TypographyFontSizeSemanticToken { get }
 
-    // MARK: - Semantic token - Typography - Font - Size - Tablet
+    // MARK: - Semantic token - Typography - Font - Size - Tablet (regular/medium
 
     var fontSizeTabletDisplayLarge: TypographyFontSizeSemanticToken { get }
     var fontSizeTabletDisplayMedium: TypographyFontSizeSemanticToken { get }
@@ -81,8 +106,6 @@ public protocol TypographySemanticTokens {
     var fontSizeTabletBodyLarge: TypographyFontSizeSemanticToken { get }
     var fontSizeTabletBodyMedium: TypographyFontSizeSemanticToken { get }
     var fontSizeTabletBodySmall: TypographyFontSizeSemanticToken { get }
-    var fontSizeTabletCodeMedium: TypographyFontSizeSemanticToken { get }
-    var fontSizeTabletCodeSmall: TypographyFontSizeSemanticToken { get }
 
     // MARK: - Semantic token - Typography - Font - Size - Others
 
@@ -93,7 +116,7 @@ public protocol TypographySemanticTokens {
     var fontSizeCodeMedium: TypographyFontSizeSemanticToken { get }
     var fontSizeCodeSmall: TypographyFontSizeSemanticToken { get }
 
-    // MARK: - Semantic token - Typography - Font - Line height - Mobile
+    // MARK: - Semantic token - Typography - Font - Line height - Mobile (extra-compact/compact)
 
     var fontLineHeightMobileDisplayLarge: TypographyFontLineHeightSemanticToken { get }
     var fontLineHeightMobileDisplayMedium: TypographyFontLineHeightSemanticToken { get }
@@ -105,10 +128,8 @@ public protocol TypographySemanticTokens {
     var fontLineHeightMobileBodyLarge: TypographyFontLineHeightSemanticToken { get }
     var fontLineHeightMobileBodyMedium: TypographyFontLineHeightSemanticToken { get }
     var fontLineHeightMobileBodySmall: TypographyFontLineHeightSemanticToken { get }
-    var fontLineHeightMobileCodeMedium: TypographyFontLineHeightSemanticToken { get }
-    var fontLineHeightMobileCodeSmall: TypographyFontLineHeightSemanticToken { get }
 
-    // MARK: - Semantic token - Typography - Font - Line height - Tablet
+    // MARK: - Semantic token - Typography - Font - Line height - Tablet (regular/medium)
 
     var fontLineHeightTabletDisplayLarge: TypographyFontLineHeightSemanticToken { get }
     var fontLineHeightTabletDisplayMedium: TypographyFontLineHeightSemanticToken { get }
@@ -120,8 +141,6 @@ public protocol TypographySemanticTokens {
     var fontLineHeightTabletBodyLarge: TypographyFontLineHeightSemanticToken { get }
     var fontLineHeightTabletBodyMedium: TypographyFontLineHeightSemanticToken { get }
     var fontLineHeightTabletBodySmall: TypographyFontLineHeightSemanticToken { get }
-    var fontLineHeightTabletCodeMedium: TypographyFontLineHeightSemanticToken { get }
-    var fontLineHeightTabletCodeSmall: TypographyFontLineHeightSemanticToken { get }
 
     // MARK: - Semantic token - Typography - Font - Line height - Others
 
@@ -135,7 +154,7 @@ public protocol TypographySemanticTokens {
     // MARK: - Semantic token - Typography - Font - Letter spacing
     // TODO: Missing details about the types of the associated raw tokens
 
-    // MARK: - Semantic token - Typography - Font - Paragraph spacing - Mobile
+    // MARK: - Semantic token - Typography - Font - Paragraph spacing - Mobile (extra-compact/compact)
 
     var fontParagraphSpacingMobileDisplayLarge: TypographyFontParagraphSpacingRawToken { get }
     var fontParagraphSpacingMobileDisplayMedium: TypographyFontParagraphSpacingRawToken { get }
@@ -147,10 +166,8 @@ public protocol TypographySemanticTokens {
     var fontParagraphSpacingMobileBodyLarge: TypographyFontParagraphSpacingRawToken { get }
     var fontParagraphSpacingMobileBodyMedium: TypographyFontParagraphSpacingRawToken { get }
     var fontParagraphSpacingMobileBodySmall: TypographyFontParagraphSpacingRawToken { get }
-    var fontParagraphSpacingMobileCodeMedium: TypographyFontParagraphSpacingRawToken { get }
-    var fontParagraphSpacingMobileCodeSmall: TypographyFontParagraphSpacingRawToken { get }
 
-    // MARK: - Semantic token - Typography - Font - Paragraph spacing - Tablet
+    // MARK: - Semantic token - Typography - Font - Paragraph spacing - Tablet (regular/medium)
 
     var fontParagraphSpacingTabletDisplayLarge: TypographyFontParagraphSpacingRawToken { get }
     var fontParagraphSpacingTabletDisplayMedium: TypographyFontParagraphSpacingRawToken { get }
@@ -162,8 +179,6 @@ public protocol TypographySemanticTokens {
     var fontParagraphSpacingTabletBodyLarge: TypographyFontParagraphSpacingRawToken { get }
     var fontParagraphSpacingTabletBodyMedium: TypographyFontParagraphSpacingRawToken { get }
     var fontParagraphSpacingTabletBodySmall: TypographyFontParagraphSpacingRawToken { get }
-    var fontParagraphSpacingTabletCodeMedium: TypographyFontParagraphSpacingRawToken { get }
-    var fontParagraphSpacingTabletCodeSmall: TypographyFontParagraphSpacingRawToken { get }
 
     // MARK: - Semantic token - Typography - Font - Paragraph spacing - Others
 
@@ -174,6 +189,41 @@ public protocol TypographySemanticTokens {
     var fontParagraphSpacingCodeMedium: TypographyFontParagraphSpacingRawToken { get }
     var fontParagraphSpacingCodeSmall: TypographyFontParagraphSpacingRawToken { get }
 
-    // MARK: - Semantic tplens - Typography - Composites
-    // TODO: Missing specifications, composite semantic tokens to dd for display, heading, body, label, code
+    // MARK: - Semantic tokens - Typography - Composites - Display
+
+    var typeDisplayLarge: TypographyCompositeSemanticToken { get }
+    var typeDisplayMedium: TypographyCompositeSemanticToken { get }
+    var typeDisplaySmall: TypographyCompositeSemanticToken { get }
+
+    // MARK: - Semantic tokens - Typography - Composites - Heading
+
+    var typeHeadingXLarge: TypographyCompositeSemanticToken { get }
+    var typeHeadingLarge: TypographyCompositeSemanticToken { get }
+    var typeHeadingMedium: TypographyCompositeSemanticToken { get }
+    var typeHeadingSmall: TypographyCompositeSemanticToken { get }
+
+    // MARK: - Semantic tokens - Typography - Composites - Body
+
+    var typeBodyDefaultLarge: TypographyCompositeSemanticToken { get }
+    var typeBodyDefaultMedium: TypographyCompositeSemanticToken { get }
+    var typeBodyDefaultSmall: TypographyCompositeSemanticToken { get }
+    var typeBodyStrongLarge: TypographyCompositeSemanticToken { get }
+    var typeBodyStrongMedium: TypographyCompositeSemanticToken { get }
+    var typeBodyStrongSmall: TypographyCompositeSemanticToken { get }
+
+    // MARK: - Semantic tokens - Typography - Composites - Label
+
+    var typeLabelDefaultXLarge: TypographyCompositeSemanticToken { get }
+    var typeLabelDefaultLarge: TypographyCompositeSemanticToken { get }
+    var typeLabelDefaultMedium: TypographyCompositeSemanticToken { get }
+    var typeLabelDefaultSmall: TypographyCompositeSemanticToken { get }
+    var typeLabelStrongXLarge: TypographyCompositeSemanticToken { get }
+    var typeLabelStrongLarge: TypographyCompositeSemanticToken { get }
+    var typeLabelStrongMedium: TypographyCompositeSemanticToken { get }
+    var typeLabelStrongSmall: TypographyCompositeSemanticToken { get }
+
+    // MARK: - Semantic tokens - Typography - Composites - Code
+
+    var typeCodeMedium: TypographyCompositeSemanticToken { get }
+    var typeCodeSmall: TypographyCompositeSemanticToken { get }
 }